New: Moved getLegendPadding() to helpers/legend.js
