+++ src/main/js/helpers/legend.js
@@ -394,6 +394,46 @@ const getDefaultLegendOptions = (graphConfig, dataTarget) => {
 
     return legendOptions;
 };
+/**
+ * Helper function to set the right legend padding values based on input JSON.
+ *
+ * @private
+ * @param {object} config - config which needs to be updated
+ * @param {object} inputLegendPadding - input legend padding provided via input JSON.
+ * @returns {object} - padding for Legend
+ */
+const getLegendPadding = (config, inputLegendPadding) => {
+    // If legendPadding is provided, update the config object with the provided values, else update it with the default constants
+    if (utils.isDefined(config.legendPadding)) {
+        return {
+            top: getDefaultValue(
+                inputLegendPadding.top,
+                constants.LEGEND_PADDING.top
+            ),
+            bottom: getDefaultValue(
+                inputLegendPadding.bottom,
+                constants.LEGEND_PADDING.bottom
+            ),
+            right: getDefaultValue(
+                inputLegendPadding.right,
+                constants.LEGEND_PADDING.right
+            ),
+            left: getDefaultValue(
+                inputLegendPadding.left,
+                constants.LEGEND_PADDING.left
+            ),
+            hasCustomLegendPadding: true
+        };
+    } else {
+        return {
+            top: constants.LEGEND_PADDING.top,
+            bottom: constants.LEGEND_PADDING.bottom,
+            right: constants.LEGEND_PADDING.right,
+            left: constants.LEGEND_PADDING.left,
+            hasCustomLegendPadding: false
+        };
+    }
+};
 
 /**
  * @enum {Function}
@@ -406,5 +446,6 @@ export {
     legendClickHandler,
     legendHoverHandler,
     isLegendSelected,
-    getDefaultLegendOptions
+    getDefaultLegendOptions,
+    getLegendPadding
 };
